backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
langchain
openai
pymongo
python-dotenv
python-jose
passlib[bcrypt]